channel-display: Avoid wrong macro expansion
The spice_channel_set_capability uses the second argument
also as a string to look for environment variable (for instance
using spice_channel_set_capability(one, two) we would look for
a "two" environment).
Usually this is correct as the macro is called with mnemonics
which looks like SPICE_xxxx_CAP_yyyy but is not for this case.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>