gstaudio: replace exisiting channels on connect
There is a race if disconnect/reconnect happen too quickly that the
existing channel isn't yet destroyed, due to the delayed coroutine
handling.

By simply taking the last channel as the current channel, we avoid the
issue having to wait for the previous channel to be destroyed.

Of course, multiple playback/record channels aren't supported anyway at
this point.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>